% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmevent.data.frame}
\alias{as.pharmevent.data.frame}
\title{Create a pharmevent table from a data.frame.}
\usage{
\method{as.pharmevent}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
pharmevent
}
\description{
Creates a pharmevent table from a data.frame.
}
\examples{
# data(pharmevents)
head(as.pharmevent(pharmevents))
}

