% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmevent.NULL}
\alias{as.pharmevent.NULL}
\title{Create an empty pharmevent table.}
\usage{
\method{as.pharmevent}{NULL}(x, ...)
}
\arguments{
\item{x}{ignored}

\item{...}{ignored}
}
\value{
pharmevent
}
\description{
Creates an empty pharmevent table. Essentially, a constructor function.
}
\examples{
as.pharmevent()
}

