% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmevent.pharmval}
\alias{as.pharmevent.pharmval}
\title{Convert pharmval to pharmevent.}
\usage{
\method{as.pharmevent}{pharmval}(x, ...)
}
\arguments{
\item{x}{pharmval object}

\item{...}{ignored}
}
\value{
pharmevent
}
\description{
Converts pharmval to pharmevent.
}
\examples{
as.pharmevent(as.pharmval())
}

