% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmevent}
\alias{as.pharmevent}
\title{Dispatch pharmevent method.}
\usage{
as.pharmevent(x = NULL, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Dispatches pharmevent method.
}

