% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmval.NULL}
\alias{as.pharmval.NULL}
\title{Create an empty pharmval table.}
\usage{
\method{as.pharmval}{NULL}(x, ...)
}
\arguments{
\item{x}{ignored}

\item{...}{ignored}
}
\value{
pharmval
}
\description{
Creates an empty pharmval table.
}
\examples{
as.pharmval()
}

