% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmval.pharmevent}
\alias{as.pharmval.pharmevent}
\title{Convert pharmevent to pharmval.}
\usage{
\method{as.pharmval}{pharmevent}(x, ...)
}
\arguments{
\item{x}{pharmevent object}

\item{...}{ignored}
}
\value{
pharmval
}
\description{
Converts pharmevent to pharmval.
}
\examples{
head(as.pharmval(as.pharmevent(pharmevents)))
}

