% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmval.R
\name{as.pharmval}
\alias{as.pharmval}
\title{Dispatch pharmval method.}
\usage{
as.pharmval(x = NULL, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Dispatches pharmval method.
}

