% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{pharmval-package}
\alias{pharmval}
\alias{pharmval-package}
\title{Integrate pharmacometric inputs, outputs, and metadata.}
\description{
\pkg{pharmval} is a dataset design that supports inputs, outputs, and associated metadata for pharmacometric analyses.  It extends the generalized dataset design of Schmidt and Radivojevic (2014) Table 3 (\url{http://dx.doi.org/10.1007/s10928-014-9370-4}).
}
\references{
Schmidt H and Radivojevic A. 2014. Enhancing population pharmacokinetic modeling efficiency and quality using an integrated workflow. Journal of Pharmacokinetics and Pharmacodynamics 41(4): 319-334.
}

